requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-redis==5.4.0
django-rq==3.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
redis==5.2.1
rq==2.1.0
six==1.17.0
sqlparse==0.5.3
supervisor==4.2.5
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
